pyproject.toml
[tool.poetry]
name = "word-bearer"
version = "0.1.0"
package-mode = false
description = ""
authors = ["adelynflowers <adelyn.flowers@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.6.4"
python-dotenv = "^1.2.1"
pydantic = "^2.12.5"
pandas = "^2.3.3"
loguru = "^0.7.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.10,<0.15.0)"
]
